+++ dimas/src/agent.rs
@@ -45,9 +45,18 @@
 //!
 
 // region:		--- modules
-use crate::context::Context;
-use crate::prelude::*;
+use crate::com::{
+	liveliness_subscriber::LivelinessSubscriberBuilder, publisher::PublisherBuilder,
+	query::QueryBuilder, queryable::QueryableBuilder, subscriber::SubscriberBuilder,
+};
+use crate::timer::TimerBuilder;
 use crate::utils::{wait_for_task_signals, TaskSignal};
+use crate::{
+	config::Config,
+	context::{ArcContext, Context},
+	error::{DimasError, Result},
+};
+use std::sync::RwLock;
 use std::{
 	fmt::Debug,
 	sync::{mpsc, Mutex},