+++ dimas/src/com/subscriber.rs
@@ -4,18 +4,27 @@
 //! A `Subscriber` can optional subscribe on a delete message.
 
 // region:		--- modules
-use crate::{prelude::*, utils::TaskSignal};
+use crate::{
+	context::ArcContext,
+	error::{DimasError, Result},
+	utils::TaskSignal,
+};
 #[allow(unused_imports)]
 use std::collections::HashMap;
-use std::sync::{mpsc::Sender, Mutex};
+use std::sync::{mpsc::Sender, Arc, Mutex};
+#[cfg(feature = "subscriber")]
+use std::sync::RwLock;
 use tokio::task::JoinHandle;
 #[cfg(feature = "subscriber")]
 use tracing::info;
 use tracing::{error, instrument, warn, Level};
 use zenoh::{
 	prelude::{r#async::AsyncResolve, SampleKind},
+	subscriber::Reliability,
 	SessionDeclarations,
 };
+
+use super::message::Message;
 // endregion:	--- modules
 
 // region:		--- types